# Performance improvement over fast iterables