--- a/src/sleep.ts
+++ b/src/sleep.ts
@@ -1,2 +1,3 @@
 import * as util from "util";
 export const sleep = util.promisify(setTimeout);
+export const breath = util.promisify(setImmediate);