Solves #23 

Some performance improvements were implemented in this PR.

* The events array now is a queue instance. A simple implementation was made in code, which will give a performance of **O(1)** for the dequeuing operation, opposed to the previous **O(N)** of destructuring an array;
* Trick deactivating and activating event listening to make waitResponse better. If waitResponse already gets a valid value to return, this value will not be included in the queue (as the listeners that do so are off) and can be yielded immediately. Just after getting the value, the original listener (that adds to the queue) is re-enabled, so there's no risk any message is lost;
* the sleep call was replaced by a promise generated from a setImmediate. This will give the same effect (making a non blocking operation), but will hang on much less, greatly increasing performance over a large amount of events;
* A no sleep option was created for the cases where the lib user doesn't care if a blocking operation is done, but cares about performance (for example, when we're creating a cli);

I also added a new test suite to validate performance, using a large file as the source of event emitting. This suite doesn't pass on the previous version, but I don't know how long it would take to finish because it timed out in 100 seconds. I think it would take 5 to 10 minutes to finish.
Using this branch locally the test finished in 386 ms with noSleep = false and 349ms with noSleep = true.

Running a separated benchmark without all the mocha overload I got the following results, but over a file with only 9074 lines:
```
before

####################################################
WARMUP: Current memory usage: 2.2 MB
WARMUP: readline forEmitOf profiling started
Read 9074 lines
WARMUP: readline forEmitOf profiling finished in 10757ms
WARMUP: Current memory usage (before GC): 15.15 MB
WARMUP: Current memory usage (after GC): 2.39 MB

####################################################
RUN 1: readline forEmitOf profiling started
Read 9074 lines
RUN 1: readline forEmitOf profiling finished in 10805ms
RUN 1: Current memory usage (before GC): 11.14 MB
RUN 1: Current memory usage (after GC): 2.39 MB

####################################################
RUN 2: readline forEmitOf profiling started
Read 9074 lines
RUN 2: readline forEmitOf profiling finished in 10798ms
RUN 2: Current memory usage (before GC): 11.91 MB
RUN 2: Current memory usage (after GC): 2.4 MB

####################################################
RUN 3: readline forEmitOf profiling started
Read 9074 lines
RUN 3: readline forEmitOf profiling finished in 10849ms
RUN 3: Current memory usage (before GC): 14.5 MB
RUN 3: Current memory usage (after GC): 2.66 MB

after

####################################################
WARMUP: Current memory usage: 2.21 MB
WARMUP: readline forEmitOf profiling started
Read 9074 lines
WARMUP: readline forEmitOf profiling finished in 56ms
WARMUP: Current memory usage (before GC): 4.21 MB
WARMUP: Current memory usage (after GC): 2.35 MB

####################################################
RUN 1: readline forEmitOf profiling started
Read 9074 lines
RUN 1: readline forEmitOf profiling finished in 47ms
RUN 1: Current memory usage (before GC): 6.75 MB
RUN 1: Current memory usage (after GC): 2.62 MB

####################################################
RUN 2: readline forEmitOf profiling started
Read 9074 lines
RUN 2: readline forEmitOf profiling finished in 52ms
RUN 2: Current memory usage (before GC): 6.32 MB
RUN 2: Current memory usage (after GC): 2.63 MB

####################################################
RUN 3: readline forEmitOf profiling started
Read 9074 lines
RUN 3: readline forEmitOf profiling finished in 46ms
RUN 3: Current memory usage (before GC): 4.8 MB
RUN 3: Current memory usage (after GC): 2.64 MB

after with noSleep

####################################################
WARMUP: Current memory usage: 2.21 MB
WARMUP: readline forEmitOf profiling started
Read 9074 lines
WARMUP: readline forEmitOf profiling finished in 23ms
WARMUP: Current memory usage (before GC): 3.39 MB
WARMUP: Current memory usage (after GC): 2.33 MB

####################################################
RUN 1: readline forEmitOf profiling started
Read 9074 lines
RUN 1: readline forEmitOf profiling finished in 16ms
RUN 1: Current memory usage (before GC): 4.38 MB
RUN 1: Current memory usage (after GC): 2.35 MB

####################################################
RUN 2: readline forEmitOf profiling started
Read 9074 lines
RUN 2: readline forEmitOf profiling finished in 14ms
RUN 2: Current memory usage (before GC): 4.18 MB
RUN 2: Current memory usage (after GC): 2.38 MB

####################################################
RUN 3: readline forEmitOf profiling started
Read 9074 lines
RUN 3: readline forEmitOf profiling finished in 16ms
RUN 3: Current memory usage (before GC): 4.21 MB
RUN 3: Current memory usage (after GC): 2.37 MB
```