@danstarns 2021-12-10
Thank you for this! Sorry, I have been distant on your raised issue #23. I will review it soon! Cheers @Farenheith 🍻